requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
APScheduler==3.10.4
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
multidict==6.0.5
PyMySQL==1.1.0
python-dotenv==1.0.1
python-telegram-bot==20.8
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tzlocal==5.2
urllib3==2.2.1
yarl==1.9.4
